requirements.txt
mysql-connector-python==8.0.29
data-diff==0.2.8
python-dotenv